requirements.txt
flask==1.0.2
gunicorn==19.9.0
google-cloud-datastore==1.7.3
grpcio==1.18.0
pandas==0.24.2
matplotlib==3.0.3
PyPDF2==1.26.0
google-cloud-bigquery
google-cloud-storage
google-cloud
